[FEAT] 일부 컴포넌트 및 훅 분리, 체결 탭 모바일
