+++ frontend/src/pages/Token/TokenDetail.tsx
@@ -72,9 +72,10 @@ export default function TokenDetail() {
           tokenOhlcv={tokenOhlcv}
           buyList={buyList}
           sellList={sellList}
+          tradeList={tradeList}
           tradePrice={selectedPrice ?? fixedPrice ?? 0}
           onPriceSelect={setSelectedPrice}
-          initialTab={location.state?.mobileTab === 'list' ? 'list' : 'chart'}
+          initialTab={location.state?.mobileTab === 'list' ? 'list' : undefined}
         />
       )}
