+++ frontend/src/pages/Token/components/mobile/MobileTokenTradeCard.tsx
@@ -1,13 +1,13 @@
-import { useState, useEffect, useCallback, useRef } from 'react';
+import { useEffect, useRef } from 'react';
 import ToggleGroup from '@/components/common/ToggleGroup';
 import Input from '@/components/common/Input';
 import Button from '@/components/common/Button';
 import PercentageBtn from '../tokenDetail/PercentageBtn';
-import { tokenApi } from '@/api/tokenApi';
 import Toast from '@/components/common/Toast';
-import type { OrderInfo, TokenOhlcv, TokenPending, Order } from '@/types/tokenType';
+import type { OrderInfo, TokenOhlcv } from '@/types/tokenType';
 import { Trashcan } from '@/components/icon/Icons';
 import { useOrderbookLadder } from '@/pages/Token/hooks/useOrderbookLadder';
+import { useTradeOrderForm } from '@/pages/Token/hooks/useTradeOrderForm';
 
 interface MobileTokenTradeCardProps {
   tokenId: number;
@@ -32,73 +32,30 @@ export default function MobileTokenTradeCard({
     { id: 'pending', label: '미체결' },
   ];
 
-  const [activeTab, setActiveTab] = useState('buy');
-  const [orderType, setOrderType] = useState('LIMIT');
-  const [price, setPrice] = useState('');
-  const [volume, setVolume] = useState('');
-  const [amount, setAmount] = useState('');
-  const [pendingList, setPendingList] = useState<TokenPending[]>([]);
-  const [toastMsg, setToastMsg] = useState<string | null>(null);
   const ladder = useOrderbookLadder({ ohlcv, buyList, sellList });
   const orderbookScrollRef = useRef<HTMLDivElement>(null);
-
-  const getNumPrice = () => Number(price.replace(/,/g, '')) || 0;
-  const getNumVolume = () => Number(volume) || 0;
-  const getNumAmount = () => Number(amount.replace(/,/g, '')) || 0;
-
-  useEffect(() => {
-    if (marketPrice && marketPrice > 0) {
-      setPrice(marketPrice.toLocaleString());
-    }
-  }, [marketPrice]);
-
-  const formatNumber = (val: string) => {
-    const num = val.replace(/[^0-9]/g, '');
-    return num ? Number(num).toLocaleString() : '';
-  };
-
-  const handleInputChange = (
-    e: React.ChangeEvent<HTMLInputElement>,
-    setter: (v: string) => void,
-  ) => {
-    const { value, selectionStart } = e.target;
-    const prevLen = value.length;
-    const formatted = formatNumber(value);
-    setter(formatted);
-
-    setTimeout(() => {
-      if (selectionStart !== null) {
-        const newPos = selectionStart + (formatted.length - prevLen);
-        e.target.setSelectionRange(newPos, newPos);
-      }
-    }, 0);
-  };
-
-  const formatDateTime = (dateString: string) => {
-    const date = new Date(dateString);
-
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, '0');
-    const day = String(date.getDate()).padStart(2, '0');
-    const hours = String(date.getHours()).padStart(2, '0');
-    const minutes = String(date.getMinutes()).padStart(2, '0');
-    const seconds = String(date.getSeconds()).padStart(2, '0');
-
-    return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-  };
-
-  const fetchPendingOrders = useCallback(async () => {
-    try {
-      const data = await tokenApi.getPendingList(tokenId);
-      setPendingList(data || []);
-    } catch (error) {
-      console.error('미체결 내역 조회 실패:', error);
-    }
-  }, [tokenId]);
-
-  useEffect(() => {
-    if (activeTab === 'pending') fetchPendingOrders();
-  }, [activeTab, fetchPendingOrders]);
+  const {
+    activeTab,
+    setActiveTab,
+    orderType,
+    setOrderType,
+    price,
+    setPrice,
+    volume,
+    setVolume,
+    amount,
+    pendingList,
+    toastMsg,
+    setToastMsg,
+    getNumPrice,
+    getNumVolume,
+    getNumAmount,
+    handleFormattedInputChange,
+    formatDateTime,
+    handlePercentageClick,
+    handleOrder,
+    cancelOrder,
+  } = useTradeOrderForm({ tokenId, marketPrice });
 
   useEffect(() => {
     if (!ohlcv || !orderbookScrollRef.current || ladder.length === 0) return;
@@ -124,106 +81,6 @@ export default function MobileTokenTradeCard({
     };
   }, [ohlcv?.tokenId, ladder.length]);
 
-  const handlePercentageClick = async (perc: number) => {
-    try {
-      let balanceStr = '0';
-      if (activeTab === 'buy') {
-        const data = await tokenApi.getCashBalance();
-        balanceStr = String(data);
-      } else {
-        const data = await tokenApi.getTokenBalance(tokenId);
-        balanceStr = String(data);
-      }
-
-      const balance = Number(balanceStr);
-      const calculatedValue = balance * (perc / 100);
-
-      if (activeTab === 'buy') {
-        if (orderType === 'MARKET') {
-          setAmount(Math.floor(calculatedValue).toLocaleString());
-        } else {
-          const currentPrice = getNumPrice();
-          if (currentPrice > 0) {
-            setVolume((calculatedValue / currentPrice).toFixed(4));
-          }
-        }
-      } else {
-        setVolume(calculatedValue.toFixed(4));
-      }
-    } catch (error) {
-      console.error('잔액 조회 실패:', error);
-    }
-  };
-
-  const handleOrder = async () => {
-    const numPrice = getNumPrice();
-    const numVolume = getNumVolume();
-    const numAmount = getNumAmount();
-
-    const order: Order = {
-      tokenId,
-      orderSide: activeTab === 'buy' ? 'BUY' : 'SELL',
-      orderType: orderType as 'LIMIT' | 'MARKET',
-      orderPrice: orderType === 'LIMIT' ? numPrice.toString() : '0',
-      orderVolume:
-        activeTab === 'buy' && orderType === 'MARKET'
-          ? '0'
-          : numVolume.toString(),
-      totalPrice:
-        activeTab === 'buy'
-          ? orderType === 'LIMIT'
-            ? (numPrice * numVolume).toString()
-            : numAmount.toString()
-          : '0',
-    };
-
-    if (order.orderSide === 'BUY') {
-      if (orderType === 'LIMIT') {
-        if (!price || price === '' || price === '0')
-          return setToastMsg('가격을 입력해주세요.');
-        if (!volume || volume === '' || volume === '0')
-          return setToastMsg('수량을 입력해주세요.');
-      } else if (orderType === 'MARKET') {
-        if (!amount || amount === '' || amount === '0')
-          return setToastMsg('주문 총액을 입력해주세요.');
-      }
-
-      if (Number(order.totalPrice) < 1000) {
-        return setToastMsg('최소 주문 금액은 1,000원입니다.');
-      }
-    } else {
-      if (!volume || volume === '' || volume === '0')
-        return setToastMsg('수량을 입력해주세요.');
-      if (orderType === 'LIMIT' && (!price || price === '' || price === '0'))
-        return setToastMsg('가격을 입력해주세요.');
-
-      if (numVolume < 0.00001)
-        return setToastMsg('최소 주문 수량은 0.00001개입니다.');
-    }
-
-    try {
-      await tokenApi.createOrder(tokenId, order);
-      setToastMsg('주문 완료');
-      setPrice('');
-      setVolume('');
-      setAmount('');
-    } catch (e) {
-      console.error('주문 실패:', e);
-      setToastMsg('주문 실패');
-    }
-  };
-
-  const cancelOrder = async (orderId: number) => {
-    try {
-      await tokenApi.cancelOrder(tokenId, orderId);
-      setToastMsg('주문 취소');
-      fetchPendingOrders();
-    } catch (e) {
-      console.error('주문 취소 실패:', e);
-      setToastMsg('주문 취소 실패');
-    }
-  };
-
   return (
     <div className="flex h-full min-h-0 overflow-hidden bg-white">
       <div className="flex w-2/5 min-w-0 flex-col border-r border-gray-100">
@@ -239,10 +96,21 @@ export default function MobileTokenTradeCard({
               <div
                 key={`${row.price}-${idx}`}
                 data-current-row={row.isCurrent ? 'true' : undefined}
-                className={`grid h-9 grid-cols-[1fr_1.3fr] ${hasBottomDivider ? 'border-b border-gray-100' : ''}`}
+                className={`grid h-9 grid-cols-[1.25fr_1fr] ${hasBottomDivider ? 'border-b border-gray-100' : ''}`}
               >
                 <div
-                  className={`flex h-full items-center justify-center px-2 text-center font-body-03 ${isSell ? 'text-info' : 'text-error'} ${priceBgClass}`}
+                  onClick={() => {
+                    setPrice(row.price.toLocaleString());
+                  }}
+                  className={`flex h-full cursor-pointer items-center justify-center px-2 text-center font-body-02 ${isSell ? 'text-info' : 'text-error'} ${priceBgClass}`}
+                  role="button"
+                  tabIndex={0}
+                  onKeyDown={(e) => {
+                    if (e.key === 'Enter' || e.key === ' ') {
+                      e.preventDefault();
+                      setPrice(row.price.toLocaleString());
+                    }
+                  }}
                   style={
                     row.isCurrent
                       ? {
@@ -316,27 +184,27 @@ export default function MobileTokenTradeCard({
                             {item.orderSide === 'BUY' ? '매수' : '매도'}
                           </span>
                         </div>
-                        <span className="font-caption-01 text-gray-400">
+                        <span className="font-caption-02 text-gray-400">
                           {formatDateTime(item.createdAt)}
                         </span>
                       </div>
                       <div className="flex flex-col gap-1.5">
                         <div className="flex justify-between">
-                          <span className="font-caption-01 text-gray-400">주문가격</span>
-                          <span className="font-caption-03">
+                          <span className="font-caption-02 text-gray-400">주문가격</span>
+                          <span className="font-body-03">
                             {Number(item.orderPrice).toLocaleString()}
                           </span>
                         </div>
                         <div className="flex justify-between">
-                          <span className="font-caption-01 text-gray-400">주문수량</span>
-                          <span className="font-caption-03">
+                          <span className="font-caption-02 text-gray-400">주문수량</span>
+                          <span className="font-body-03">
                             {Number(item.orderVolume).toFixed(4)}
                           </span>
                         </div>
                         <div className="flex justify-between">
-                          <span className="font-caption-01 text-gray-400">미체결량</span>
+                          <span className="font-caption-02 text-gray-400">미체결량</span>
                           <span
-                            className={`font-caption-03 ${item.orderSide === 'BUY' ? 'text-error' : 'text-info'}`}
+                            className={`font-body-03 ${item.orderSide === 'BUY' ? 'text-error' : 'text-info'}`}
                           >
                             {Number(item.remainingToken).toFixed(4)}
                           </span>
@@ -355,11 +223,11 @@ export default function MobileTokenTradeCard({
         ) : (
           <div className="flex min-h-0 flex-1 flex-col gap-4 pt-4">
             <div className="flex items-center gap-2">
-              <span className="w-24 text-gray-400 font-caption-02">주문 유형</span>
+              <span className="w-24 text-gray-400 font-caption-03">주문 유형</span>
               <select
                 value={orderType}
                 onChange={(e) => setOrderType(e.target.value)}
-                className="w-full h-full min-h-[42px] px-2 pr-8 bg-gray-50 border border-gray-100 rounded-[var(--radius-s)] outline-none font-caption-02 text-gray-900 cursor-pointer hover:border-gray-200 transition-colors py-0 leading-[40px] flex items-center"
+                className="w-full h-full min-h-[42px] px-2 pr-8 bg-gray-50 border border-gray-100 rounded-[var(--radius-s)] outline-none font-caption-01 text-gray-900 cursor-pointer hover:border-gray-200 transition-colors py-0 leading-[40px] flex items-center"
               >
                 <option value="LIMIT">지정가</option>
                 <option value="MARKET">시장가</option>
@@ -368,13 +236,13 @@ export default function MobileTokenTradeCard({
 
             {orderType === 'LIMIT' && (
               <div className="flex items-center gap-2">
-                <span className="w-24 text-gray-400 font-caption-02">
+                <span className="w-24 text-gray-400 font-caption-03">
                   {activeTab === 'buy' ? '매수 가격' : '매도 가격'}
                 </span>
                 <Input
                   height={42}
                   value={price}
-                  onChange={(e) => handleInputChange(e, setPrice)}
+                  onChange={(e) => handleFormattedInputChange(e, 'price')}
                   placeholder="예: 100,000"
                 />
               </div>
@@ -382,17 +250,17 @@ export default function MobileTokenTradeCard({
 
             {activeTab === 'buy' && orderType === 'MARKET' ? (
               <div className="flex items-center gap-2">
-                <span className="w-24 text-gray-400 font-caption-02">주문 금액</span>
+                <span className="w-24 text-gray-400 font-body-03">주문 금액</span>
                 <Input
                   height={42}
                   value={amount}
-                  onChange={(e) => handleInputChange(e, setAmount)}
+                  onChange={(e) => handleFormattedInputChange(e, 'amount')}
                   placeholder="예: 100,000"
                 />
               </div>
             ) : (
               <div className="flex items-center gap-2">
-                <span className="w-24 text-gray-400 font-caption-02">주문 수량</span>
+                <span className="w-24 text-gray-400 font-caption-03">주문 수량</span>
                 <Input
                   height={42}
                   value={volume}
@@ -409,7 +277,7 @@ export default function MobileTokenTradeCard({
             />
 
             <div className="mt-auto flex items-center justify-between pt-4">
-              <span className="text-gray-900 font-body-03">
+              <span className="text-gray-900 font-body-04">
                 총 {activeTab === 'buy' ? '주문 금액' : '주문 수량'}
               </span>
               <span
